Small hotfix in deployment docs due to typo in previous PR.
Summary: This revision is a hotfix, I noticed that the end of the deployment docs had an incomplete phrase which was probably a typo in this PR: https://github.com/dagster-io/dagster/commit/c9bc657c6ed2cab2569120c68fe30879a88ed4de. This fixes it for posterity.

Test Plan: unit

Reviewers: max, prha

Reviewed By: max

Differential Revision: https://dagster.phacility.com/D1558