ML Pipeline Outline
Summary:
**Summary**

The following PR set's up the scaffolding of our first machine learning pipeline. This is by no means intended to be a good model with state of the art techniques, but should be an accurate reflection of what ml pipeline code would look like. I opted to use Keras just because it was super duper easy to get a quick and dirty model up. Pytorch takes a bit configuring to do right.

**Changelog**

- Tweaked training set to transpose matrix to work well with Keras tensors.
- Built training solid with config setups.
- Edited unit tests to reflect these changes.

Test Plan: unit

Reviewers: alangenfeld, max, nate

Reviewed By: nate

Differential Revision: https://dagster.phacility.com/D1546